move ContentSeparator and BtnContainer to common
